feat(testchain-gen): allow setting storage path and network

+++ ledger/testchain-generator/src/main.rs
@@ -13,32 +13,46 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use snarkvm_console::prelude::{Network as _, TestRng};
+use snarkvm_console::prelude::{CanaryV0, MainnetV0, Network, TestRng, TestnetV0};
 use snarkvm_ledger::{Ledger, store::helpers::rocksdb::ConsensusDB, test_helpers::TestChainBuilder};
 
 use aleo_std::StorageMode;
 
-use anyhow::{Context, Result};
-use clap::Parser;
-
-type Network = snarkvm_console::prelude::TestnetV0;
+use anyhow::{Context, Result, bail};
+use clap::{Parser, builder::PossibleValuesParser};
 
 #[derive(Parser)]
 struct Args {
+    /// The number of validators active on the chain.
+    /// (also corresponds to the number of certificates per round)
     num_validators: usize,
+    /// The number of blocks to generator.
     num_blocks: usize,
+    /// Load the genesis block for the chain from the specified path instead of generating it.
     #[clap(long)]
     genesis_path: Option<String>,
+    /// Set a custom storage path for the ledger.
+    /// By default, it will use the ledger of the first devnet validator.
+    #[clap(long, short = 'p')]
+    storage_path: Option<String>,
+    /// Remove existing ledger if it already exists.
+    #[clap(long, short = 'f')]
+    force: bool,
+    /// The name of the network to generate the chain for.
+    #[clap(long, value_parser=PossibleValuesParser::new(vec![CanaryV0::SHORT_NAME, TestnetV0::SHORT_NAME, MainnetV0::SHORT_NAME]), default_value=TestnetV0::SHORT_NAME)]
+    network: String,
 }
 
 /// Removes an existing ledger (if any) from the filesystem.
-fn remove_ledger(network: u16) -> Result<()> {
-    let path = aleo_std::aleo_ledger_dir(network, &StorageMode::Development(0));
+fn remove_ledger(network: u16, storage_mode: &StorageMode, force: bool) -> Result<()> {
+    let path = aleo_std::aleo_ledger_dir(network, storage_mode);
 
-    if path.exists() {
+    if path.exists() && force {
         std::fs::remove_dir_all(&path).with_context(|| "Failed to remove existing ledger")?;
 
-        println!("Remove existing ledger data at {path:?}");
+        println!("Removed existing ledger data at {path:?}");
+    } else if path.exists() {
+        bail!("There is already a ledger at {path:?}. Re-run with `--force` if you want to overwrite it");
     }
 
     Ok(())
@@ -50,22 +64,31 @@ fn main() -> Result<()> {
 
     let args = Args::parse();
 
-    let mut rng = TestRng::default();
+    match args.network.as_str() {
+        CanaryV0::SHORT_NAME => generate_testchain::<CanaryV0>(args),
+        TestnetV0::SHORT_NAME => generate_testchain::<TestnetV0>(args),
+        MainnetV0::SHORT_NAME => generate_testchain::<MainnetV0>(args),
+        // This is caught by clap.
+        _ => unreachable!(),
+    }
+}
 
-    remove_ledger(Network::ID)?;
+fn generate_testchain<N: Network>(args: Args) -> Result<()> {
+    let mut rng = TestRng::default();
+    let storage_mode =
+        if let Some(path) = args.storage_path { StorageMode::Custom(path.into()) } else { StorageMode::Development(0) };
+
+    remove_ledger(N::ID, &storage_mode, args.force)?;
 
     let num_validators = args.num_validators;
     let num_blocks = args.num_blocks;
 
     println!("Initializing test chain builder with {num_validators} validators");
-    let mut builder = match args.genesis_path {
-        Some(genesis_path) => {
-            TestChainBuilder::<Network>::new_with_quorum_size_and_genesis_block(num_validators, genesis_path)
-                .with_context(|| "Failed to set up test chain builder")?
-        }
-        None => TestChainBuilder::<Network>::new_with_quorum_size(num_validators, &mut rng)
-            .with_context(|| "Failed to set up test chain builder")?,
-    };
+    let mut builder: TestChainBuilder<N> = match args.genesis_path {
+        Some(genesis_path) => TestChainBuilder::new_with_quorum_size_and_genesis_block(num_validators, genesis_path),
+        None => TestChainBuilder::new_with_quorum_size(num_validators, &mut rng),
+    }
+    .with_context(|| "Failed to set up test chain builder")?;
 
     println!("Generating {num_blocks} blocks");
 
@@ -82,9 +105,8 @@ fn main() -> Result<()> {
     }
 
     println!("Done. Storing blocks to disk.");
-    let ledger =
-        Ledger::<Network, ConsensusDB<Network>>::load(builder.genesis_block().clone(), StorageMode::Development(0))
-            .with_context(|| "Failed to initialize ledger")?;
+    let ledger = Ledger::<N, ConsensusDB<N>>::load(builder.genesis_block().clone(), storage_mode)
+        .with_context(|| "Failed to initialize ledger")?;
 
     // Ensure there is only one active ledger at a time.
     drop(builder);
